requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.25.0

# Document Processing (with fallbacks)
PyPDF2>=3.0.1

# AI/ML - Core (CPU versions for Render)
sentence-transformers>=2.2.2
torch>=2.0.0,<2.1.0
transformers>=4.35.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Vector Store (optional)
pinecone-client>=3.0.0

# Async file handling
aiofiles>=23.0.0

# Optional document processing
python-docx>=1.1.0
pdfplumber>=0.10.3

# Security (optional)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0